Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main()
     int a,b,c;
     float Average;
     scanf("%a %b %c",&a,&b,&c);
-    Average =(float)(a+b+c)/3;
+    Average =(a+b+c)/3;
     printf("Average: %.2f",Average);
     return 0;
 }
